+++ src/functions.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use Bakame\Stackwatch\AggregatorType;
+use Bakame\Stackwatch\AggregationType;
 use Bakame\Stackwatch\Metrics;
 use Bakame\Stackwatch\Report;
 use Bakame\Stackwatch\Result;
@@ -13,10 +13,12 @@ if (!function_exists('stack')) {
      * Profile a callable, returns a @see Result object.
      *
      * The function returns the Result object generated for further usage if needed.
+     *
+     * @param ?non-empty-string $label
      */
-    function stack(callable $callback): Result
+    function stack(callable $callback, ?string $label = null): Result
     {
-        return Stack::execute($callback);
+        return Stack::execute($callback, $label);
     }
 }
 
@@ -25,10 +27,12 @@ if (!function_exists('stack_dump')) {
      * Profile a callable, dump the generated Span and return the callback returned value.
      *
      * The function returns the Result object generated for further usage if needed.
+     *
+     * @param ?non-empty-string $label
      */
-    function stack_dump(callable $callback): Result
+    function stack_dump(callable $callback, ?string $label = null): Result
     {
-        $result = Stack::execute($callback);
+        $result = Stack::execute($callback, $label);
         $result->span->dump();
 
         return $result;
@@ -40,10 +44,12 @@ if (!function_exists('stack_dd')) {
      * Profile a callable, dump the generated Span and return the callback returned value.
      *
      * The function returns the callable expecged value
+     *
+     * @param ?non-empty-string $label
      */
-    function stack_dd(callable $callback): never
+    function stack_dd(callable $callback, ?string $label = null): never
     {
-        Stack::execute($callback)->span->dd();
+        Stack::execute($callback, $label)->span->dd();
     }
 }
 
@@ -56,7 +62,7 @@ if (!function_exists('stack_report')) {
      *
      * @throws Throwable
      */
-    function stack_report(callable $callback, int $iterations = 3, int $warmup = 0): Report
+    function stack_report(callable $callback, int $iterations = 1, int $warmup = 0): Report
     {
         return Stack::report($callback, $iterations, $warmup);
     }
@@ -73,7 +79,7 @@ if (!function_exists('stack_rdump')) {
      *
      * @throws Throwable
      */
-    function stack_rdump(callable $callback, int $iterations = 3, int $warmup = 0): Report
+    function stack_rdump(callable $callback, int $iterations = 1, int $warmup = 0): Report
     {
         return Stack::dumpReport($callback, $iterations, $warmup);
     }
@@ -88,7 +94,7 @@ if (!function_exists('stack_rdd')) {
      *
      * @throws Throwable
      */
-    function stack_rdd(callable $callback, int $iterations = 3, int $warmup = 0): never
+    function stack_rdd(callable $callback, int $iterations = 1, int $warmup = 0): never
     {
         Stack::ddReport($callback, $iterations, $warmup);
     }
@@ -103,7 +109,7 @@ if (!function_exists('stack_metrics')) {
      *
      * @throws Throwable
      */
-    function stack_metrics(callable $callback, int $iterations = 3, int $warmup = 0, AggregatorType $type = AggregatorType::Average): Metrics
+    function stack_metrics(callable $callback, int $iterations = 1, int $warmup = 0, AggregationType $type = AggregationType::Average): Metrics
     {
         return Stack::metrics($callback, $iterations, $warmup, $type);
     }
@@ -120,7 +126,7 @@ if (!function_exists('stack_mdump')) {
      *
      * @throws Throwable
      */
-    function stack_mdump(callable $callback, int $iterations = 3, int $warmup = 0, AggregatorType $type = AggregatorType::Average): Metrics
+    function stack_mdump(callable $callback, int $iterations = 1, int $warmup = 0, AggregationType $type = AggregationType::Average): Metrics
     {
         return Stack::dumpMetrics($callback, $iterations, $warmup, $type);
     }
@@ -135,7 +141,7 @@ if (!function_exists('stack_mdd')) {
      *
      * @throws Throwable
      */
-    function stack_mdd(callable $callback, int $iterations = 3, int $warmup = 0, AggregatorType $type = AggregatorType::Average): never
+    function stack_mdd(callable $callback, int $iterations = 1, int $warmup = 0, AggregationType $type = AggregationType::Average): never
     {
         Stack::ddMetrics($callback, $iterations, $warmup, $type);
     }